+++ CHANGELOG.md
@@ -1,6 +1,10 @@
-## Changelog for 0.10.1
+## Changelog for 0.10.2
 
 ### Improvements
 
-- **Chat Enhancements**
-  - Added copy message with right-click.
+- **RRV Ingestion**
+  - Improved search calc
+  - Hide Skyblock quick-nav buttons when in the recipe viewer
+  - Fixed crash
+  - Maybe fixed performance issues
+  - Added custom button textures for categories. Huge thanks to [Bentcheesee](https://modrinth.com/user/Bentcheesee) again.